Current limiting

+++ mode.c
@@ -3,21 +3,7 @@
  *
  * Author: Tuomas Vaherkoski <tuomasvaherkoski@gmail.com>
  *
- * This file is part of variable-power-supply-oshw-project.
- *
- * This program free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
+ * This file is part of variable-power-supply project.
  */
 
 #include "display.h"
@@ -48,7 +34,7 @@ void return_previous_mode(void) {
 /* limits */
 uint16_t current_limit; // mA
 
-void set_current_limit(unsigned int limit) {
+void set_current_limit(uint16_t limit) {
     if(limit >= 10 && limit < 3000) {
         current_limit = limit;
     } else if (limit < 10) {
